move multi option component into common
